Add new optional user_identifier_claims bundle config
This allows the user to define which claim in the token is used for the
user ID. I couldn't find a good way to figure out how to automate this,
so leave it to the user. Consider the following cases:

* In case Keycloak is the sole IDP and also used for all user specific
  information then using the UUID provided via the "sub" claim is probably
  what is wanted.
* In case Keycloak is just federating LDAP users then other systems are
  likely also using LDAP, and the LDAP cn is probably mapped to some
  claim.
* In case Keycloak also contains service accounts which the system wants
  to handle as "users", then using "azp" as a fallback might be wanted,
  to give those users the client name as user ID in case no other user
  name claim is present.

To handle all those cases introduce "user_identifier_claims" which is a list
of claims which will be tried one after the other until one is available in the
token. If none of them is found then the user still gets a "null" user ID.

The default settings match what we had before.