OIDCUserSessionProvider: provide a better session ID for service accounts
At least in combination with Keycloak. For service accounts Keycloak doesn't
set the session_state claim which we use for deriving a session ID. In that
case fall back to the token ID, so at least all requests made with the same
token show up with the same session ID in the logs. Not perfect, but better
than before.