AuthorizationDataProvider: skip in case the userid doesn't match the token
This normally doesn't happen since we always use this in the context of an
authenticated user and always only ask for the current user, but in case
of health checks we ask unauthenticated.

Skip using any token derived info in case the token isn't set, the userId is
missing, or in case the userId doesn't match the one derived from the token.

This should make sure that we only use the token info in case the current user
is authenticated and we ask for attributes of that specific user.

Adjust the tests to always pass a userId, to keep them working as before.