config: add "set_symfony_roles_from_scopes" option for disable role mapping
By default this bundle will use the token scopes and map them to Symfony roles
for the authenticated user.

We are slowly moving away from using Symfony roles to attribute based authz,
so add an option so we can disable this feature.

Setting it to false means the user will have no custom Symfony roles when
authenticated.